Refactor GUI app module into dedicated submodules
